+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Smart field - Enable search on smart fields.
 
 ## RELEASE 1.3.44 - 2016-10-11
 ### Fixed